feat: Added test mode
New mandatory argument passed with -T is test directory name.
If arg is missing, first we check if we're on amboso >1.1.2+, if so we first try for "./kazoj", then we read from stego.lock.

Test mode should ATM override -prib, as it will make amboso quit before reaching any of those stages.
For now, the actual test routine is just a mock up recording the two available tests.

kazoj.lock houses dirnames for the case folder and the errorcase folder.